Astromactch finalizado
