--- quinzena8/banu-astromatch/src/NOVO/components/Footer/index.js
@@ -1,6 +1,0 @@
-function Footer(params) {
-    return (
-        <div>Footer</div>
-    )
-}
-export default Footer;
