--- quinzena8/banu-astromatch/src/NOVO/components/Header/style.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-export const ContanierHeader = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    border: 1px solid black;
-    width: 370px;
-`
-export const AstroMatchLogo = '../img/logo_astromatch.png'
