--- quinzena8/banu-astromatch/src/NOVO/components/Home/index.js
@@ -1,6 +1,0 @@
-function Home(params) {
-    return (
-        <div>Home</div>
-    )
-}
-export default Home;
