--- quinzena8/banu-astromatch/src/NOVO/components/Matchs/index.js
@@ -1,6 +1,0 @@
-function Matchs(params) {
-    return (
-        <div>Matchs</div>
-    )
-}
-export default Matchs;
