--- quinzena8/banu-astromatch/src/NOVO/components/Profile/index.js
@@ -1,6 +1,0 @@
-function Profile(params) {
-    return (
-        <div>Profine</div>
-    )
-}
-export default Profile;
