--- quinzena8/banu-astromatch/src/NOVO/constants/url.js
@@ -1,2 +1,0 @@
-export const urlAstroMatch = 'https://us-central1-missao-newton.cloudfunctions.net/'
-
