+++ quinzena8/banu-astromatch/src/components/Main/style.js
@@ -5,14 +5,14 @@ export const ContanierMain = styled.div`
     flex-direction: column;
     justify-content: flex-start;
     align-items: center;
-    border: 1px solid black;
-    width: 370px;
-    height: 660px;
+    background-color: #48cdee;
+    max-width: 370px;
+    max-height: 600px;
+    min-height: 450px;
 `
 export const ContainerPrincipal = styled.div`
     display: flex;
     flex-direction: column;
     align-items: center;
     width: 100%;
-    background-color: red;
 `
