+++ quinzena8/banu-astromatch/src/constants/url.js
@@ -1,3 +1,2 @@
 export const urlAstroMatch = 'https://us-central1-missao-newton.cloudfunctions.net/'
-export const ImgCard = 'https://conteudo.imguol.com.br/c/entretenimento/a1/2020/06/04/qual-e-a-receita-da-quarentena-de-manu-gavassi-descubra-e-experimente-1591301290665_v2_450x600.jpg'
