+++ quinzena8/banu-astromatch/src/index.css
@@ -6,7 +6,6 @@
 body {
   margin: 0;
   padding: 0;
-  
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
